chore(config): use FilteredConfig for IgnoredMatch
