+++ ggshield/core/config/user_config.py
@@ -1,11 +1,10 @@
 import os
-from dataclasses import dataclass, field, fields
+from dataclasses import dataclass, field
 from typing import Any, Dict, Iterable, List, Optional, Set, Tuple
 
 import click
 import marshmallow_dataclass
 from marshmallow import ValidationError
-from marshmallow.decorators import pre_load
 
 from ggshield.core.config.errors import ParseError, format_validation_error
 from ggshield.core.config.utils import (
@@ -20,32 +19,12 @@ from ggshield.core.constants import (
     GLOBAL_CONFIG_FILENAMES,
     LOCAL_CONFIG_PATHS,
 )
-from ggshield.core.text_utils import display_warning
-from ggshield.core.types import IgnoredMatch, IgnoredMatchSchema
+from ggshield.core.types import FilteredConfig, IgnoredMatch, IgnoredMatchSchema
 from ggshield.core.utils import api_to_dashboard_url
 from ggshield.iac.utils import POLICY_ID_PATTERN, validate_policy_id
 
 
 CURRENT_CONFIG_VERSION = 2
-
-
-@marshmallow_dataclass.dataclass
-class FilteredConfig:
-    @classmethod
-    @pre_load(pass_many=False)
-    def filter_fields(cls, data: Dict, **kwargs: Any) -> Dict:
-        """
-        Remove and alert on unknown fields.
-        """
-        field_names = {field_.name for field_ in fields(cls)}
-        filtered_fields = {}
-        for key, item in data.items():
-            if key in field_names:
-                filtered_fields[key] = item
-            else:
-                display_warning("Unrecognized key in config: {}".format(key))
-
-        return filtered_fields
 
 
 @marshmallow_dataclass.dataclass
@@ -244,7 +223,6 @@ class UserV1Config:
         ignored_matches = [
             ignored_match_schema.load(secret) for secret in v1config.matches_ignore
         ]
-
         secret = SecretConfig(
             show_secrets=v1config.show_secrets,
             ignored_detectors=v1config.banlisted_detectors,
